requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
astroid==3.3.10
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
crewai==0.1.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==4.4.2
dill==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
ffmpeg-python==0.2.0
fonttools==4.58.0
frozenlist==1.6.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
langchain==0.0.335
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langsmith==0.0.87
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
mccabe==0.7.0
moviepy==1.0.3
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.40.0
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
proglog==0.1.12
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pylint==3.3.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
streamlit==1.45.1
tenacity==8.5.0
tiktoken==0.5.2
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
yarl==1.20.0
zipp==3.21.0
